feat: move questionnaireAccordion to common
