+++ src/components/app/pagePoliticianDetails/ca/index.tsx
@@ -1,7 +1,7 @@
 import { QUESTIONNAIRE_HASH_KEY } from '@/components/app/pagePoliticianDetails/common/constants'
 import { PagePoliticianDetails } from '@/components/app/pagePoliticianDetails/common/politiciansDetails'
+import { QuestionnaireAccordion } from '@/components/app/pagePoliticianDetails/common/questionnaireAccordion'
 import { ScoreExplainer } from '@/components/app/pagePoliticianDetails/common/scoreExplainer'
-import { QuestionnaireAccordion } from '@/components/app/pagePoliticianDetails/us/questionnaireAccordion'
 import { ScrollToTopOnRender } from '@/components/app/scrollToTopOnRender'
 import { DTSIPersonDetails } from '@/data/dtsi/queries/queryDTSIPersonDetails'
 import { NormalizedQuestionnaire } from '@/utils/server/builder/models/data/questionnaire'
